Moved getWebDriverInstance() to inc-lib-test.php.
