+++ tests/selenium_tests/phpunit_bootstrap.php
@@ -71,29 +71,3 @@ define('WEBDRIVER_IMPLICIT_WAIT', 30);
 // Time to wait when no expected condition can be set (in seconds)
 define('SELENIUM_TEST_SLEEP', 10);
 
-function getWebDriverInstance()
-{
-    // Provide a re-usable webdriver for selenium tests.
-
-    global $_INI;
-    static $webDriver;
-
-    if (!isset($webDriver)) {
-        // Create Firefox webdriver
-        $capabilities = array(WebDriverCapabilityType::BROWSER_NAME => 'firefox');
-        $webDriver = RemoteWebDriver::create('http://localhost:4444/wd/hub', $capabilities,
-                                             WEBDRIVER_MAX_WAIT_DEFAULT * 1000,
-                                             WEBDRIVER_MAX_WAIT_DEFAULT * 1000);
-
-        // Set time for trying to access DOM elements
-        $webDriver->manage()->timeouts()->implicitlyWait(WEBDRIVER_IMPLICIT_WAIT);
-
-        if (isset($_INI['test']['xdebug_enabled']) && $_INI['test']['xdebug_enabled'] == 'true') {
-            // Enable remote debugging by setting XDebug session cookie.
-            $webDriver->manage()->addCookie(array(
-                'name' => 'XDEBUG_SESSION',
-                'value' => 'selenium'));
-        }
-    }
-    return $webDriver;
-}
